Split PackIndex impls into multiple files
